pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "weather-trading"
version = "0.4.42"
description = "Foundations for a Polymarket weather trading system."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0",
    "httpx>=0.28",
    "SQLAlchemy>=2.0",
    "aiosqlite>=0.20",
    "pandas>=2.2",
    "numpy>=2.0",
    "eth-account>=0.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.6",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
